requirements.txt
# Core LangChain & LangGraph
langchain>=0.1.0
langgraph>=0.0.20
langchain-anthropic>=0.1.1

# Google Drive API
google-api-python-client>=2.108.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Slack Bot
slack-bolt>=1.18.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0

# Parsing
beautifulsoup4>=4.12.2

# API Server
fastapi
uvicorn

# Agent dependencies
anthropic
google-generativeai